modular is linkable

+++ esp32_arm/src/CNC/Actuator/stepper/mechanic.cpp
@@ -1,6 +1,6 @@
 #include "mechanic.h"
 
 
-uint32_t ActuatorMechanicStepper::GetStepsPerCncUnit(){
+float ActuatorMechanicStepper::GetStepsPerCncUnit(){
 
 }
