+++ esp32_arm/src/CNC/Actuator/stepper/mechanic_polor.h
@@ -5,7 +5,7 @@
 
 class ActuatorMechanic_StepperPolor: public ActuatorMechanicStepper{
     public:
-        uint32_t GetStepsPerCncUnit() override;
+        float GetStepsPerCncUnit() override;
 
 
     protected:
